feat(pypi): fallback to virtualenv if venv fails
On Ubuntu 22.04 and the `python3-venv` package is not installed:

* Importing the `venv` module will succeed.
* Running `python3 -m venv --help` will output help.
* Running `python3 -m venv <VENV_DIR>` fails.

When running in a rootless environment, i.e. not able to install the
`python3-venv` package, is not tenable. However, the `virtualenv`
package can be installed via `pip` in the users home directory.

This patch checks the result of `python3 -m venv `<VENV_DIR>`, if that
fails then attempt to call `python3 -m virtualenv <VENV_DIR>` instead.

* [ ] The error messages for this are pretty unhelpful, how could they
  be improved?